CHATBOTV2/chatbot-generator/chatbot-template/chatbot-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pandas